+++ Models/Permission.cs
@@ -19,7 +19,7 @@ namespace Models
             this.Users = new HashSet<User>();
         }
     
-        public decimal ID { get; set; }
+        public int ID { get; set; }
         public string Name { get; set; }
     
         public virtual ICollection<User> Users { get; set; }
